code-examples/chapter-19/progressive/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
pyjwt>=2.8.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-multipart>=0.0.6

